Add files via upload
For my Collaborative Science Project, I was tasked to create a window that allowed the user to input the values they had previously gotten (Such as angle and adjacent), to find the height of a mangrove tree. Further along, two entries, light intensity and the height of the mangrove tree, can be inputted by the user to be entered into the database and create a graph to see the correlation between the two variables.